fix: all dependencies fixed
