+++ src/js/script.js
@@ -1,7 +1,7 @@
 import createCategories from './categories.js';
-import { resetGame } from './game.js';
+import { resetGame, checkToggle, gameToggle } from './game.js';
 import createScore, { findWord } from './score.js';
-import checkToggle, { gameToggle } from './toggle.js';
+// import checkToggle, { gameToggle } from './toggle.js';
 import createItems, { arrayOfcards } from './arrays.js';
 
 const cardsGrid = document.querySelector('.cards');
@@ -30,7 +30,7 @@ cardsGrid.addEventListener('click', (e) => {
     const index = Number(e.target.parentNode.parentNode.getAttribute('data')) + 1;
     createItems(index);
     gameToggle.checked = false;
-    resetGame();
+    // resetGame();
     checkToggle();
     const menuItems = document.querySelectorAll('.nav-link');
     menuItems.forEach((item) => {
@@ -57,21 +57,17 @@ offcanvasBody.addEventListener('click', (e) => {
     document.querySelectorAll('.active').forEach((item) => item.classList.remove('active'));
     e.target.classList.add('active');
     const categoryId = Number(e.target.getAttribute('data'));
+    resetGame();
+    gameToggle.checked = false;
+    checkToggle();
     if (categoryId < 0) {
-      resetGame();
       createScore();
     }
     if (categoryId > 0) {
       createItems(categoryId);
-      gameToggle.checked = false;
-      resetGame();
-      checkToggle();
     }
     if (categoryId === 0) {
       createCategories();
-      gameToggle.checked = false;
-      resetGame();
-      checkToggle();
     }
   }
 });
