+++ src/scss/app.scss
@@ -158,6 +158,7 @@ th {
 
 .reset-repeat-plate {
     width: 100%;
+    margin-bottom: 16px;
 }
 
 .star {
